+++ autobarrel.json
@@ -2,7 +2,8 @@
   "prefix": "/* eslint-disable padding-line-between-statements */",
   "paths": [
     "packages/graphql/src/**/*",
-    "packages/data-context/src/**/*"
+    "packages/data-context/src/**/*",
+    "packages/scaffold-config/src/**"
   ],
   "ignore": [
     "packages/data-context/src/gen",
