+++ packages/scaffold-config/src/index.ts
@@ -1,5 +1,7 @@
-export * from './configFiles'
-
-export * from './types'
+/* eslint-disable padding-line-between-statements */
+// created by autobarrel, do not modify directly
 
 export * from './constants'
+export * from './detect'
+export * from './frameworks'
+export * from './types'
