+++ packages/scaffold-config/test/unit/detect.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
 import { e2eProjectDirs } from '@packages/frontend-shared/cypress/e2e/support/e2eProjectDirs'
-import { detect } from '../../src/configFiles'
+import { detect } from '../../src'
 import Fixtures from '@tooling/system-tests/lib/fixtures'
 import path from 'path'
 import fs from 'fs-extra'
@@ -10,7 +10,7 @@ export async function scaffoldMigrationProject (project: typeof e2eProjectDirs[n
 
   await Fixtures.scaffoldProject(project)
 
-  return fs.readJSON(path.join(Fixtures.projectPath(project), 'package.json'))
+  return fs.readJSONSync(path.join(Fixtures.projectPath(project), 'package.json'))
 }
 
 describe('detect', () => {
