--- packages/scaffold-config/test/unit/scaffold-config.spec.ts
@@ -1,7 +1,0 @@
-import { expect } from 'chai'
-
-describe('@packages/scaffold-config unit', () => {
-  it('has a sample test', () => {
-    expect(1).to.eq(1)
-  })
-})
